add check autentikasi dan voter u/ pertanyaan

+++ app/Http/Controllers/PertanyaanController.php
@@ -7,9 +7,20 @@ use App\Pertanyaan;
 use App\Jawaban;
 use App\Pkomentar;
 use App\Jkomentar;
+use App\Vote;
 
 class PertanyaanController extends Controller
 {
+        /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except'=>['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -49,6 +60,7 @@ class PertanyaanController extends Controller
         $pertanyaan->judul_pertanyaan = $request->input('judul');
         $pertanyaan->isi_pertanyaan = $request->input('isi_pertanyaan');
         $pertanyaan->tag = $request->input('tag');
+        $pertanyaan->user_id = auth()->user()->id;
         $pertanyaan->save();
 
         return redirect('/pertanyaan')->with('success', 'Article created');
@@ -64,8 +76,9 @@ class PertanyaanController extends Controller
     public function show($id)
     {
         $judul =  Pertanyaan::find($id);
-        // dd(Pertanyaan::find($id));
-        return \view('UI.show')->with('pertanyaan', $judul);
+        $user = auth()->user();
+        // $comment = Pertanyaan::find($id);
+        return \view('UI.show')->with('pertanyaan', $judul)->with('user', $user);
     }
 
     /**
@@ -77,6 +90,11 @@ class PertanyaanController extends Controller
     public function edit($id)
     {
         $judul =  Pertanyaan::find($id);
+        
+
+        if (auth()->user()->id !== $judul->user_id) {
+            return \redirect('/pertanyaan')->with('error', 'Unauthorize');
+        }
         return \view('UI.edit')->with('pertanyaan', $judul);
     }
 
@@ -113,6 +131,9 @@ class PertanyaanController extends Controller
     public function destroy($id)
     {
         $article = Pertanyaan::find($id);
+        if (auth()->user()->id !== $article->user_id) {
+            return \redirect('/pertanyaan')->with('error', 'Unauthorize');
+        }
         $article->delete();
         return \redirect('/pertanyaan')->with('success', 'Question Deleted');
     }
